+++ src/froGH/froGH/Data/SelfOrganizingMapComponent.cs
@@ -4,7 +4,6 @@ using Grasshopper;
 using Grasshopper.Kernel;
 using Grasshopper.Kernel.Data;
 using Grasshopper.Kernel.Types;
-using Rhino.Geometry;
 using System;
 
 namespace froGH