+++ src/froGH/froGH/Geometry/PointsInAABBRTree.cs
@@ -8,14 +8,14 @@ using System.Collections.Generic;
 
 namespace froGH
 {
-    public class PointsInSphereRTree : GH_Component
+    public class PointsInAABBRTree : GH_Component
     {
         /// <summary>
-        /// Initializes a new instance of the PointsInSphereRTree class.
+        /// Initializes a new instance of the PointsInAABBRTree class.
         /// </summary>
-        public PointsInSphereRTree()
-          : base("Points In Sphere (RTree)", "f_PtsSphRT",
-              "Detects Points within a Sphere of given center and radius in a given RTree\nUse this for fixed search sets\nFor dynamic search sets use the other Points In Sphere",
+        public PointsInAABBRTree()
+          : base("Points In AABB (RTree)", "f_PtsAABBRT",
+              "Detects Points within a Bounding Box in a given RTree\nUse this for fixed search sets\nFor dynamic search sets use the other Points In AABB",
               "froGH", "Geometry")
         {
         }
@@ -25,8 +25,7 @@ namespace froGH
         /// </summary>
         protected override void RegisterInputParams(GH_Component.GH_InputParamManager pManager)
         {
-            pManager.AddPointParameter("Sphere Center", "P", "Point to search from (needle)", GH_ParamAccess.tree);
-            pManager.AddNumberParameter("Search Radius", "R", "Search Radius", GH_ParamAccess.tree);
+            pManager.AddBoxParameter("Bounding Box", "B", "The World XYZ aligned Bounding Box for Search", GH_ParamAccess.tree);
             pManager.AddGenericParameter("RTree", "RT", "The RTree to search", GH_ParamAccess.tree);
         }
 
@@ -35,7 +34,7 @@ namespace froGH
         /// </summary>
         protected override void RegisterOutputParams(GH_Component.GH_OutputParamManager pManager)
         {
-            pManager.AddPointParameter("Points", "P", "The Points within the sphere radius", GH_ParamAccess.tree);
+            pManager.AddPointParameter("Points", "P", "The Points within the Bounding Box", GH_ParamAccess.tree);
             pManager.AddIntegerParameter("Points indices", "i", "Indices of the points in the original list", GH_ParamAccess.tree);
         }
 
@@ -50,48 +49,40 @@ namespace froGH
             // I already had a component with dynamic RTree search by Sphere, it only made sense
             // to me to complete the set. Plus, I had the chance to study and learn some new tricks!
 
-            GH_Structure<GH_Point> points = new GH_Structure<GH_Point>();
-            GH_Structure<GH_Number> radiuses = new GH_Structure<GH_Number>();
+            GH_Structure<GH_Box> boxes = new GH_Structure<GH_Box>();
             GH_Structure<IGH_Goo> fTrees = new GH_Structure<IGH_Goo>();
 
-            if (!DA.GetDataTree(0, out points)) return;
-            if (!DA.GetDataTree(1, out radiuses)) return;
-            if (!DA.GetDataTree(2, out fTrees)) return;
+            if (!DA.GetDataTree(0, out boxes)) return;
+            if (!DA.GetDataTree(1, out fTrees)) return;
 
-            if (points == null) return;
-            if (radiuses == null) return;
+            if (boxes == null) return;
             if (fTrees == null) return;
 
             GH_Structure<GH_Point> foundPts = new GH_Structure<GH_Point>();
             GH_Structure<GH_Integer> foundInd = new GH_Structure<GH_Integer>();
 
-            for (int i = 0; i < points.PathCount; i++)
+            for (int i = 0; i < boxes.PathCount; i++)
             {
-                GH_Path path = points.Paths[i];
+                GH_Path path = boxes.Paths[i];
                 List<IGH_Goo> fTreeList = fTrees.Branches[i < fTrees.PathCount - 1 ? i : fTrees.PathCount - 1];
-                // old condition - problematic when fTrees.PathCount = 1 and points.PathCount > 1
-                //List<IGH_Goo> fTreeList = fTrees.Branches[fTrees.PathCount < i ? fTrees.PathCount - 1 : i];
                 if (fTreeList.Count == 0)
                 {
-                    AddRuntimeMessage(GH_RuntimeMessageLevel.Error, "No RTree for branch" + points.Paths[i].ToString());
+                    AddRuntimeMessage(GH_RuntimeMessageLevel.Error, "No RTree for branch" + boxes.Paths[i].ToString());
                     return;
                 }
                 if (fTreeList.Count > 1)
                 {
-                    AddRuntimeMessage(GH_RuntimeMessageLevel.Warning, "Multiple RTrees for branch" + points.Paths[i].ToString() + "to search multiple RTrees place them in individual branches");
+                    AddRuntimeMessage(GH_RuntimeMessageLevel.Warning, "Multiple RTrees for branch" + boxes.Paths[i].ToString() + "to search multiple RTrees place them in individual branches");
                     return;
                 }
                 FroGHRTree fTree = ((GH_froGHRTree)fTreeList[0]).Value;
-                List<GH_Number> radList = radiuses.Branches[i < radiuses.PathCount - 1 ? i : radiuses.PathCount - 1];
-                // old condition - problematic when radiuses.PathCount = 1 and points.PathCount > 1
-                //List<GH_Number> radList = radiuses.Branches[radiuses.PathCount < i ? radiuses.PathCount - 1 : i];
                 int count = 0;
-                foreach (GH_Point ghP in points.Branches[i])
+                foreach (GH_Box ghB in boxes.Branches[i])
                 {
-                    if (ghP == null) continue;
+                    if (ghB == null) continue;
                     GH_Path foundPath = path.AppendElement(count);
                     // perform search
-                    List<int> fIndices = fTree.IndicesInSphere(ghP.Value, radList[radList.Count > count ? count : radList.Count - 1].Value);
+                    List<int> fIndices = fTree.IndicesInAABB(ghB.Value.BoundingBox);
                     if (fIndices.Count == 0)
                     {
                         foundPts.AppendRange(new List<GH_Point>(), foundPath);
@@ -123,7 +114,7 @@ namespace froGH
             {
                 //You can add image files to your project resources and access them like this:
                 // return Resources.IconForThisComponent;
-                return Resources.PointsInSphereRTree_GH;
+                return Resources.PointsInAABBRTree_GH;
             }
         }
 
@@ -132,7 +123,7 @@ namespace froGH
         /// </summary>
         public override Guid ComponentGuid
         {
-            get { return new Guid("95705848-2386-41cd-8c50-e1e56c9fd0e4"); }
+            get { return new Guid("FD56D4D2-59F7-4B8C-BBED-67215C5C67C7"); }
         }
     }
 }