+++ src/froGH/froGH/Mesh/MeshIsocurvesScalar.cs
@@ -9,7 +9,7 @@ using System.Threading.Tasks;
 
 namespace froGH
 {
-    public class MeshIsocurves : GH_Component
+    public class MeshIsocurvesScalar : GH_Component
     {
         // constants and LUTs
         // Quads
@@ -25,15 +25,12 @@ namespace froGH
         // table of cases for Tri faces (1 to 6, case 0 and 7 leave the face empty)
         private readonly int[][] casesTri = new int[][] { new int[] { 1, 2 }, new int[] { 0, 1 }, new int[] { 0, 2 }, new int[] { 0, 2 }, new int[] { 0, 1 }, new int[] { 1, 2 } };
 
-        // precalc values
-        private double[,] normVValues;
-
         /// <summary>
         /// Initializes a new instance of the MeshIsocurves class.
         /// </summary>
-        public MeshIsocurves()
-          : base("Mesh Isocurves", "f_MIso",
-              "Generates isocurves on colored meshes",
+        public MeshIsocurvesScalar()
+          : base("Mesh Isocurves Scalar", "f_MIsoSc",
+              "Generates isocurves on Mesh from scalar values",
               "froGH", "Mesh")
         {
         }
@@ -44,7 +41,7 @@ namespace froGH
         protected override void RegisterInputParams(GH_Component.GH_InputParamManager pManager)
         {
             pManager.AddMeshParameter("Mesh", "M", "The Colored Mesh", GH_ParamAccess.item);
-            pManager.AddIntegerParameter("Color Channel", "C", "The Color channel to sample\n0 - Red\n1 - Green\n2 - Blue", GH_ParamAccess.item, 0);
+            pManager.AddNumberParameter("Mesh Scalar", "S", "Scalar values for the vertices.", GH_ParamAccess.list);
             pManager.AddNumberParameter("Isovalue", "i", "The isovalue to search for (0-1)", GH_ParamAccess.item, 0.5);
         }
 
@@ -65,83 +62,55 @@ namespace froGH
             Mesh M = new Mesh();
             if (!DA.GetData(0, ref M)) return;
             if (!(M.IsValid) || M == null) return;
-            int ch = 0;
-            if (!DA.GetData(1, ref ch)) return;
+            List<double> scalarValues = new List<double>();
+            if (!DA.GetDataList(1, scalarValues)) return;
             double iso = 0;
             if (!DA.GetData(2, ref iso)) return;
 
-            // if Mesh has no colors or if not all faces are quads throw an error
-            if (M.VertexColors.Count == 0)
+            // if scalar values do not match number of vertices throw an error
+            if (M.Vertices.Count != scalarValues.Count)
             {
-                AddRuntimeMessage(GH_RuntimeMessageLevel.Error, "Mesh must have vertex colors");
+                AddRuntimeMessage(GH_RuntimeMessageLevel.Error, "Number of values does not match number of Mesh vertices");
                 return;
             }
 
-            Precalc(M, ch);
+            double[] normalizedValues = NormalizeValues(scalarValues);
 
-            List<Line> isoLines = ToLineList(IsoCurvesOnColoredMesh(M, ch, iso));
+            List<Line> isoLines = ToLineList(IsoCurvesOnMesh(M, normalizedValues, iso));
 
             DA.SetDataList(0, isoLines);
         }
 
-        private void Precalc(Mesh M, int channel)
+        double[] NormalizeValues(List<double> values)
         {
-            normVValues = new double[M.Vertices.Count,3];
+            double[] normalizedValues = new double[values.Count];
+            List<double> sortedValues = new List<double>(values);
 
-            // compile vertices values and status
-            Parallel.For(0, M.Vertices.Count, i =>
+            // find values domain
+            sortedValues.Sort();
+
+            double denominator = 1 / (sortedValues[sortedValues.Count - 1] - sortedValues[0]);
+            // convert values
+            //for (int i = 0; i < values.Count; i++)
+            Parallel.For(0, values.Count, i =>
             {
-                // store normalized color values per vertex
-                Color c = M.VertexColors[i];
-                switch (channel)
-                {
-                    case 0:
-                        normVValues[i,0] = c.R / 255.0;
-                        break;
-                    case 1:
-                        normVValues[i,1] = c.G / 255.0;
-                        break;
-                    case 2:
-                        normVValues[i,2] = c.B / 255.0;
-                        break;
-                    default:
-                        goto case 1;
-                }
+                normalizedValues[i] = ((values[i] - sortedValues[0]) * denominator);
             });
+
+            return normalizedValues;
         }
 
-        Line[][] IsoCurvesOnColoredMesh(Mesh M, int channel, double iso)
+        Line[][] IsoCurvesOnMesh(Mesh M, double[] normVValues, double iso)
         {
             // Variables
             int[] isoStatus = new int[M.Vertices.Count];
-            //double[] normVValues = new double[M.Vertices.Count];
 
-
-            // compile vertices values and status
+            // find vertex iso status (0 = below or equal to iso, 1 = above iso)
             Parallel.For(0, M.Vertices.Count, i =>
             {
-                // store normalized color values per vertex
-                //Color c = M.VertexColors[i];
-                //switch (channel)
-                //{
-                //    case 0:
-                //        normVValues[i] = c.R / 255.0;
-                //        break;
-                //    case 1:
-                //        normVValues[i] = c.G / 255.0;
-                //        break;
-                //    case 2:
-                //        normVValues[i] = c.B / 255.0;
-                //        break;
-                //    default:
-                //        goto case 1;
-                //}
-                // find vertex iso status (0 = below or equal to iso, 1 = above iso)
-                isoStatus[i] = Convert.ToInt32(normVValues[i,channel] > iso);
-
+                isoStatus[i] = Convert.ToInt32(normVValues[i] > iso);
             });
 
-            //DataTree<Line> LineSegments = new DataTree<Line>();
             Line[][] LineSegmentsArray = new Line[M.Faces.Count][];
 
             // Performing algorithm
@@ -173,16 +142,16 @@ namespace froGH
                             // find value in face center as corner values average
                             double centerValue = 0;
                             for (int j = 0; j < 4; j++)
-                                centerValue += normVValues[ind[j], channel];
+                                centerValue += normVValues[ind[j]];
 
                             centerValue *= 0.25;
                             int centercase = Convert.ToInt32(centerValue > iso);
 
                             Point3d a, b, c, d;
-                            a = LerpPoint(M.Vertices[ind[0]], M.Vertices[ind[1]], normVValues[ind[0], channel], normVValues[ind[1], channel], iso);
-                            b = LerpPoint(M.Vertices[ind[1]], M.Vertices[ind[2]], normVValues[ind[1], channel], normVValues[ind[2], channel], iso);
-                            c = LerpPoint(M.Vertices[ind[2]], M.Vertices[ind[3]], normVValues[ind[2], channel], normVValues[ind[3], channel], iso);
-                            d = LerpPoint(M.Vertices[ind[3]], M.Vertices[ind[0]], normVValues[ind[3], channel], normVValues[ind[0], channel], iso);
+                            a = LerpPoint(M.Vertices[ind[0]], M.Vertices[ind[1]], normVValues[ind[0]], normVValues[ind[1]], iso);
+                            b = LerpPoint(M.Vertices[ind[1]], M.Vertices[ind[2]], normVValues[ind[1]], normVValues[ind[2]], iso);
+                            c = LerpPoint(M.Vertices[ind[2]], M.Vertices[ind[3]], normVValues[ind[2]], normVValues[ind[3]], iso);
+                            d = LerpPoint(M.Vertices[ind[3]], M.Vertices[ind[0]], normVValues[ind[3]], normVValues[ind[0]], iso);
 
                             if ((caseIndex == 5 && centercase == 1) || (caseIndex == 10 && centercase == 0))
                             {
@@ -207,7 +176,7 @@ namespace froGH
                             {
                                 indA = ind[edgeIndexes[j]];
                                 indB = ind[(edgeIndexes[j] + 1) % 4];
-                                lineEnds[j] = LerpPoint(M.Vertices[indA], M.Vertices[indB], normVValues[indA, channel], normVValues[indB, channel], iso);
+                                lineEnds[j] = LerpPoint(M.Vertices[indA], M.Vertices[indB], normVValues[indA], normVValues[indB], iso);
                             }
 
                             LineSegmentsArray[i] = new Line[1];
@@ -230,7 +199,7 @@ namespace froGH
                         caseIndex += vPowersTri[j] * isoStatus[ind[j]];
 
                     // excluding null cases
-                    if (caseIndex > 0 && caseIndex < 7) 
+                    if (caseIndex > 0 && caseIndex < 7)
                     {
                         int[] edgeIndexes = casesTri[caseIndex - 1];
                         Point3d[] lineEnds = new Point3d[2];
@@ -239,7 +208,7 @@ namespace froGH
                         {
                             indA = ind[edgeIndexes[j]];
                             indB = ind[(edgeIndexes[j] + 1) % 3];
-                            lineEnds[j] = LerpPoint(M.Vertices[indA], M.Vertices[indB], normVValues[indA, channel], normVValues[indB, channel], iso);
+                            lineEnds[j] = LerpPoint(M.Vertices[indA], M.Vertices[indB], normVValues[indA], normVValues[indB], iso);
                         }
 
                         LineSegmentsArray[i] = new Line[1];
@@ -300,7 +269,7 @@ namespace froGH
             {
                 //You can add image files to your project resources and access them like this:
                 // return Resources.IconForThisComponent;
-                return Resources.MeshIsocurves_GH;
+                return Resources.MeshIsocurvesScalar_GH;
             }
         }
 
@@ -309,7 +278,7 @@ namespace froGH
         /// </summary>
         public override Guid ComponentGuid
         {
-            get { return new Guid("6DDA029F-593E-4F63-B30B-5241396C3F45"); }
+            get { return new Guid("2C46F59B-DAD2-4F74-82F6-355B8861F11F"); }
         }
     }
 }