+++ src/froGH/froGH/Utils/KDNode.cs
@@ -1,0 +1,15 @@
+﻿using Rhino.Geometry;
+
+namespace froGH.Utils
+{
+    public class KDNode
+    {
+        public Point3d pt;
+        public KDNode left, right;
+
+        public KDNode(Point3d point)
+        {
+            pt = point;
+        }
+    }
+}